Remove locks around erasures

+++ implementation/rocksdb/db/zns_impl/table/iterators/sstable_iterator_compressed.h
@@ -16,7 +16,7 @@ namespace ROCKSDB_NAMESPACE {
 class SSTableIteratorCompressed : public Iterator {
  public:
   SSTableIteratorCompressed(const Comparator* comparator, char* data,
-                            uint32_t data_size, uint32_t num_restarts);
+                            uint64_t data_size, uint64_t num_restarts);
   ~SSTableIteratorCompressed();
   bool Valid() const override {
     return current_ < data_size_ && current_ >= kv_pairs_offset_ &&
@@ -45,21 +45,21 @@ class SSTableIteratorCompressed : public Iterator {
     return comparator_->Compare(a, b);
   }
   // Return the offset in data_ just past the end of the current entry.
-  inline uint32_t NextEntryOffset() const {
+  inline uint64_t NextEntryOffset() const {
     return (value_.data() + value_.size()) - data_;
   }
-  uint32_t GetRestartPoint(uint32_t index);
-  void SeekToRestartPoint(uint32_t index);
+  uint64_t GetRestartPoint(uint64_t index);
+  void SeekToRestartPoint(uint64_t index);
 
   const Comparator* const comparator_;
   char* data_;                   // underlying block contents
-  uint32_t const num_restarts_;  // Number of uint32_t entries in restart array
-  const uint32_t kv_pairs_offset_;
+  uint64_t const num_restarts_;  // Number of uint32_t entries in restart array
+  const uint64_t kv_pairs_offset_;
 
   // current_ is offset in data_ of current entry.  >= restarts_ if !Valid
-  uint32_t current_;
-  uint32_t restart_index_;  // Index of restart block in which current_ falls
-  uint32_t data_size_;      // size of data array
+  uint64_t current_;
+  uint64_t restart_index_;  // Index of restart block in which current_ falls
+  uint64_t data_size_;      // size of data array
   std::string key_;
   Slice value_;
   Status status_;
